api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
openai==1.57.0
supabase==2.10.0
python-multipart==0.0.18
httpx==0.27.2
pydantic==2.10.3
pydantic-settings==2.6.1
tiktoken==0.8.0
